Add files via upload
Excited to share my portfolio as a milestone in my internship journey! 🚀
🌐 Project Title: [Portfolio]
💼 Internship at: CodersCave
🔨 Tech Stack: HTML, CSS, JavaScript

During this internship, I've had the opportunity to work on some amazing projects, and I'm proud to showcase the skills and knowledge I've gained. My portfolio is a collection of my best work, reflecting my growth and commitment to becoming a top-notch coder.